tests/wpt/tests/encoding-detection/is-windows-1252.tentative.html

<!doctype html>
<title>is windows-1252</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Þetta er kóðunarpróf á staf. Fyrir sum tungumál sem nota latneska stafi þurfum við meira inntak til að taka ákvörðunina.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1252", 'Expected windows-1252');
    }, "Check detection result");
      done();
};
</script>
